Use of []int instead of []uint64, []int64 for shape and stride.

+++ accessors_test.go
@@ -58,7 +58,7 @@ func TestShape(t *testing.T) {
 		sz := rnd()
 		a = NewArray64(nil, sz...)
 		for i, v := range a.Shape() {
-			if a.shape[i] != uint64(v) {
+			if a.shape[i] != v {
 				t.Log("Change at", i, "was", a.shape[i], "is", v)
 				t.Fail()
 			}
@@ -67,11 +67,11 @@ func TestShape(t *testing.T) {
 		sh := a.Shape()
 		sh[ch]--
 		for i, v := range a.shape {
-			if uint64(sh[i]) != v && i != ch {
+			if sh[i] != v && i != ch {
 				t.Log("Change at", i, "was", a.shape[i], "is", v)
 				t.Fail()
 			}
-			if uint64(sh[i]) == v && i == ch {
+			if sh[i] == v && i == ch {
 				t.Log("Change propagated at", i, "was", a.shape[i], "is", v)
 				t.Fail()
 			}
@@ -113,9 +113,9 @@ func TestAt(t *testing.T) {
 	for i := 0; i < 5; i++ {
 		for j := 0; j < 5; j++ {
 			for k := 0; k < 5; k++ {
-				if a.at([]uint64{uint64(i), uint64(j), uint64(k)}) != a.At(i, j, k) {
+				if a.at([]int{i, j, k}) != a.At(i, j, k) {
 					t.Log("at failed for index", i, j, k)
-					t.Log(a.at([]uint64{uint64(i), uint64(j), uint64(k)}), "!=", a.At(i, j, k))
+					t.Log(a.at([]int{i, j, k}), "!=", a.At(i, j, k))
 					t.Fail()
 				}
 			}
